remove redundant toDouble

+++ src/main/java/at/hannibal2/skyhanni/features/misc/items/EstimatedItemValueCalculator.kt
@@ -886,7 +886,7 @@ object EstimatedItemValueCalculator {
                 val ingredient = PrimitiveIngredient(ingredients)
 
                 totalPrice += if (ingredient.isCoin()) {
-                    ingredient.count.toDouble()
+                    ingredient.count
                 } else {
                     ingredient.internalName.getPrice() * ingredient.count
                 }
